+++ spec/features/transfers/users_can_edit_project_details_spec.rb
@@ -254,4 +254,39 @@ RSpec.feature "Users can edit transfer project details" do
       expect(page).to have_content("No")
     end
   end
+
+  scenario "outgoing trust close once this transfer is completed" do
+    visit project_information_path(project)
+
+    row = find("#outgoingTrustDetails .govuk-summary-list__row:nth-child(7)")
+
+    within(row) do
+      expect(page).to have_content("No")
+      click_link "Change"
+    end
+
+    within("#outgoing-trust-to-close") do
+      choose("Yes")
+    end
+
+    click_on "Continue"
+
+    within(row) do
+      expect(page).to have_content("Yes")
+    end
+
+    within(row) do
+      click_link "Change"
+    end
+
+    within("#outgoing-trust-to-close") do
+      choose("No")
+    end
+
+    click_on "Continue"
+
+    within(row) do
+      expect(page).to have_content("No")
+    end
+  end
 end