+++ CHANGELOG.md
@@ -29,6 +29,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   project.
 - Users can now change the due to financial, safeguarding or governance issues
   response for a transfer project.
+- Users can now change the outgoing trust close once this transfer is completed
+  response for a transfer project.
 
 ## [Release-62][release-62]